requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
cryptography==42.0.7
jinja2==3.1.4
python-dotenv==1.0.1
bcrypt==3.2.2
